requirements.txt
click
gql
plotly
pyjwt
aiohttp
pandas
numpy